requirements.txt
langchain
python-dotenv
langchain-openai
langchain-core
langchain-community
bs4
faiss-cpu
from_root
langchain_groq
transformers
wikipedia
arxiv
langgraph
IPython
langgraph-cli[inmem]
streamlit
graphviz